+++ domain/telemetry.go
@@ -70,6 +70,16 @@ var (
 	// * quote - the quote asset symbol
 	SQSPricingFallbackCounterMetricName = "sqs_pricing_fallback_total"
 
+	// sqs_passthrough_numia_aprs_fetch_error_total
+	//
+	// counter that measures the number of errors when fetching APRs from Numia in a passthrough module.
+	SQSPassthroughNumiaAPRsFetchErrorCounterMetricName = "sqs_passthrough_numia_aprs_fetch_error_total"
+
+	// sqs_passthrough_timeseries_pool_fees_fetch_error_total
+	//
+	// counter that measures the number of errors when fetching fees from timeseries data stack in a passthrough module.
+	SQSPassthroughTimeseriesPoolFeesFetchErrorCounterMetricName = "sqs_passthrough_timeseries_pool_fees_fetch_error_total"
+
 	SQSIngestHandlerProcessBlockDurationGauge = prometheus.NewGauge(
 		prometheus.GaugeOpts{
 			Name: SQSIngestUsecaseProcessBlockDurationMetricName,
@@ -137,6 +147,20 @@ var (
 		},
 		[]string{"base", "quote"},
 	)
+
+	SQSPassthroughNumiaAPRsFetchErrorCounter = prometheus.NewCounter(
+		prometheus.CounterOpts{
+			Name: SQSPassthroughNumiaAPRsFetchErrorCounterMetricName,
+			Help: "Total number of errors when fetching APRs from Numia in a passthrough module.",
+		},
+	)
+
+	SQSPassthroughTimeseriesPoolFeesFetchErrorCounter = prometheus.NewCounter(
+		prometheus.CounterOpts{
+			Name: SQSPassthroughTimeseriesPoolFeesFetchErrorCounterMetricName,
+			Help: "Total number of errors when fetching pool fees from timeseries in a passthrough module.",
+		},
+	)
 )
 
 func init() {
@@ -149,4 +173,6 @@ func init() {
 	prometheus.MustRegister(SQSUpdateAssetsAtHeightIntervalErrorCounter)
 	prometheus.MustRegister(SQSPricingErrorCounter)
 	prometheus.MustRegister(SQSPricingFallbackCounter)
+	prometheus.MustRegister(SQSPassthroughNumiaAPRsFetchErrorCounter)
+	prometheus.MustRegister(SQSPassthroughTimeseriesPoolFeesFetchErrorCounter)
 }