+++ passthrough/usecase/passthrough_usecase.go
@@ -14,6 +14,7 @@ import (
 	"github.com/osmosis-labs/sqs/domain/mvc"
 	passthroughdomain "github.com/osmosis-labs/sqs/domain/passthrough"
 	"github.com/osmosis-labs/sqs/log"
+	"github.com/osmosis-labs/sqs/sqsutil/datafetchers"
 )
 
 type passthroughUseCase struct {
@@ -23,6 +24,10 @@ type passthroughUseCase struct {
 	defaultQuoteDenom     string
 	liquidityPricer       domain.LiquidityPricer
 	passthroughGRPCClient passthroughdomain.PassthroughGRPCClient
+
+	aprPrefetcher datafetchers.Fetcher[map[uint64]passthroughdomain.PoolAPR]
+
+	poolFeesPrefetcher datafetchers.Fetcher[map[uint64]passthroughdomain.PoolFee]
 
 	logger log.Logger
 }
@@ -95,6 +100,11 @@ const (
 	// 2. Concentrated positions
 	pooledBalancedNumJobs = 2
 
+	// Number of unclaimed rewards jobs to fetch concurrently.
+	// 1. Unclaimed rewards from concentrated positions
+	// 2. Unclaimed rewards from staking rewards
+	unclaimedRewardsNumJobs = 2
+
 	// locked + unlocking
 	numInLocksQueries = 2
 )
@@ -126,7 +136,7 @@ func (p *passthroughUseCase) GetPortfolioAssets(ctx context.Context, address str
 	defer close(pooledBalancesChan)
 
 	// Channel to fetch unclaimed rewards concurrently.
-	unclaimedRewardsChan := make(chan coinsResult)
+	unclaimedRewardsChan := make(chan coinsResult, unclaimedRewardsNumJobs)
 	defer close(unclaimedRewardsChan)
 
 	go func() {
@@ -159,6 +169,17 @@ func (p *passthroughUseCase) GetPortfolioAssets(ctx context.Context, address str
 		// Send unclaimed rewards to the unclaimed rewards channel
 		unclaimedRewardsChan <- coinsResult{
 			coins: unclaimedRewads,
+			err:   err,
+		}
+	}()
+
+	go func() {
+		// Fetch unclaimed staking rewards concurrently
+		unclaimedStakingRewards, err := p.passthroughGRPCClient.DelegationRewards(ctx, address)
+
+		// Send unclaimed rewards to the unclaimed rewards channel
+		unclaimedRewardsChan <- coinsResult{
+			coins: unclaimedStakingRewards,
 			err:   err,
 		}
 	}()
@@ -197,8 +218,29 @@ func (p *passthroughUseCase) GetPortfolioAssets(ctx context.Context, address str
 
 	// Callback to fetch unclaimed rewards concurrently.
 	getUnclaimedRewards := func(ctx context.Context, address string) (sdk.Coins, error) {
-		unclaimedRewardsResult := <-unclaimedRewardsChan
-		return unclaimedRewardsResult.coins, unclaimedRewardsResult.err
+		unclaimedCoins := sdk.Coins{}
+
+		var finalErr error
+		for i := 0; i < unclaimedRewardsNumJobs; i++ {
+			unclaimedRewardsResult := <-unclaimedRewardsChan
+
+			if unclaimedRewardsResult.err != nil {
+				// Rather than returning the error, log it and continue
+				finalErr = unclaimedRewardsResult.err
+
+				// Ensure that coins are valid to be added and avoid panic.
+				if len(unclaimedRewardsResult.coins) > 0 && !unclaimedRewardsResult.coins.IsAnyNil() {
+					unclaimedCoins = unclaimedCoins.Add(unclaimedRewardsResult.coins...)
+				}
+
+				continue
+			}
+
+			unclaimedCoins = unclaimedCoins.Add(unclaimedRewardsResult.coins...)
+		}
+
+		// Return error and best-effort result
+		return unclaimedCoins, finalErr
 	}
 
 	// Fetch jobs to fetch the portfolio assets concurrently in separate gorooutines.
@@ -348,6 +390,16 @@ func (p *passthroughUseCase) GetPortfolioAssets(ctx context.Context, address str
 	}
 
 	return finalResult, nil
+}
+
+// RegisterAPRFetcher registers the APR fetcher for the passthrough use case.
+func (p *passthroughUseCase) RegisterAPRFetcher(aprFetcher datafetchers.Fetcher[map[uint64]passthroughdomain.PoolAPR]) {
+	p.aprPrefetcher = aprFetcher
+}
+
+// RegisterPoolFeesFetcher registers the pool fees fetcher for the passthrough use case.
+func (p *passthroughUseCase) RegisterPoolFeesFetcher(poolFeesFetcher datafetchers.Fetcher[map[uint64]passthroughdomain.PoolFee]) {
+	p.poolFeesPrefetcher = poolFeesFetcher
 }
 
 // computeCapitalizationForCoins instruments the coins with their liquiditiy capitalization values.