+++ sqsutil/datafetchers/prefetch.go
@@ -6,6 +6,13 @@ import (
 	"time"
 )
 
+// Fetcher is an interface that provides a method to get a value.
+type Fetcher[T any] interface {
+	Get() (T, time.Time, error)
+	GetRefetchInterval() time.Duration
+	WaitUntilFirstResult()
+}
+
 // IntervalFetcher is a struct that prefetches a value at a given interval
 // and provides a method to get the latest value.
 // NOTE: It may return stale data if the update function takes longer than the interval.
@@ -94,4 +101,8 @@ func (p *IntervalFetcher[T]) Close() {
 
 	p.hasClosed = true
 	p.timer.Stop()
+}
+
+func (p *IntervalFetcher[T]) GetRefetchInterval() time.Duration {
+	return p.interval
 }